Sieve of Eratosthenes module creation
Sieve of Eratosthenes module created to allow escalability. Primos.py now only contains the function call within the loop.
